Update PACs for all devices

+++ pac/atsame54p20a/src/sercom0/i2cm.rs
@@ -107,7 +107,7 @@ impl crate::Writable for ADDR {}
 #[doc = "I2CM Address"]
 pub mod addr;
 #[doc = "I2CM Data\n\nThis register you can [`read`](crate::generic::Reg::read), [`reset`](crate::generic::Reg::reset), [`write`](crate::generic::Reg::write), [`write_with_zero`](crate::generic::Reg::write_with_zero), [`modify`](crate::generic::Reg::modify). See [API](https://docs.rs/svd2rust/#read--modify--write-api).\n\nFor information about available fields see [data](data) module"]
-pub type DATA = crate::Reg<u32, _DATA>;
+pub type DATA = crate::Reg<u8, _DATA>;
 #[allow(missing_docs)]
 #[doc(hidden)]
 pub struct _DATA;
